+++ web/documentserver-example/java-spring/src/main/resources/static/scripts/jscript.js
@@ -16,10 +16,23 @@
  *
  */
 
+var directUrl;
+
 if (typeof jQuery !== "undefined") {
     jq = jQuery.noConflict();
 
+    directUrl = getUrlVars()["directUrl"] == "true";
+
     mustReload = false;
+
+    if (directUrl)
+        jq("#directUrl").prop("checked", directUrl);
+    else
+        directUrl = jq("#directUrl").prop("checked");
+
+    jq("#directUrl").change(function() {
+        window.location = "?directUrl=" + jq(this).prop("checked");
+    });
 
     jq(function () {
         jq("#fileupload").fileupload({
@@ -228,7 +241,7 @@ if (typeof jQuery !== "undefined") {
 
     jq(document).on("click", "#beginEdit:not(.disable)", function () {
         var fileId = encodeURIComponent(jq("#hiddenFileName").val());
-        var url = UrlEditor + "?action=edit&fileName=" + fileId;
+        var url = UrlEditor + "?action=edit&fileName=" + fileId + "&directUrl=" + directUrl;
         window.open(url, "_blank");
         jq("#hiddenFileName").val("");
         jq.unblockUI();
@@ -236,7 +249,7 @@ if (typeof jQuery !== "undefined") {
 
     jq(document).on("click", "#beginView:not(.disable)", function () {
         var fileId = encodeURIComponent(jq("#hiddenFileName").val());
-        var url = UrlEditor + "?action=view&fileName=" + fileId;
+        var url = UrlEditor + "?action=view&fileName=" + fileId + "&directUrl=" + directUrl;
         window.open(url, "_blank");
         jq("#hiddenFileName").val("");
         jq.unblockUI();
@@ -244,7 +257,7 @@ if (typeof jQuery !== "undefined") {
 
     jq(document).on("click", "#beginEmbedded:not(.disable)", function () {
         var fileId = encodeURIComponent(jq("#hiddenFileName").val());
-        var url = UrlEditor + "?type=embedded&action=embedded&fileName=" + fileId;
+        var url = UrlEditor + "?type=embedded&action=embedded&fileName=" + fileId + "&directUrl=" + directUrl;
 
         jq("#mainProgress").addClass("embedded");
         jq("#beginEmbedded").addClass("disable");
@@ -298,6 +311,17 @@ if (typeof jQuery !== "undefined") {
             jq("div#portal-info").hide();
             jq("div.stored-list").show();
         }
+    };
+
+    function getUrlVars() {
+        var vars = [], hash;
+        var hashes = window.location.href.slice(window.location.href.indexOf('?') + 1).split('&');
+        for (var i = 0; i < hashes.length; i++) {
+            hash = hashes[i].split('=');
+            vars.push(hash[0]);
+            vars[hash[0]] = hash[1];
+        }
+        return vars;
     };
 
     jq("#portal-info")[0].innerHTML += jq("#portal-info")[0].attributes.tooltip.value;