+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Change Log
 
+- java spring: option to send directUrl
 - nodejs: referenceData
 
 ## 1.4.0
@@ -14,6 +15,7 @@
 - changed jwt implementation in csharp, csharp-mvc, php, ruby
 
 ## 1.2.0
+
 - ruby v3.0
 - set filetype in setHistoryData
 - read filetype from input request
@@ -29,6 +31,7 @@
 - ignore certificate
 
 ## 1.1.0
+
 - creating docxf
 - opening docxf, oform